Fixing reading and comparing
Still not quite satisfied with how read was accomplished.
Really looking forward to find a better solution because with this fix now the return value of fread() is always 1 because the last line will be always read through.
One solution could be to flush the input on the file pointer just skipping the newline with an fseek. Though not sure it will resolve the fread() return problem.