Implementing remove command
It is now possible to remove an author directly from terminal.
The remotion locates only completely matching strings: the limitation are those of the insertion.
Like insertion, right now the remove action fails and closes if a buffer of length 0 is passed.

README must be updated.
Docs should be updated.